test: extend engine property timeout
